+++ src/components/partners/partner/Partner.component.jsx
