+++ src/components/partners/partners-list/PartnersList.component.jsx
