+++ src/components/partners/partners-list/partners_list.styles.scss
