+++ src/components/team-member/team_member.styles.scss
